Ai_code/pyproject.toml
[project]
name = "mod-audio-stream-bridge"
version = "0.2.0"
description = "OpenAI Realtime <-> FreeSWITCH bridge for mod_audio_stream with custom voice support"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "websockets>=12.0",
  "python-dotenv>=1.0.0",
  "aiohttp>=3.9.0",
]

[project.optional-dependencies]
hq = [
  "soxr>=0.3.0",
  "numpy>=1.24.0",
]
alt = [
  "samplerate>=0.2.0",
  "numpy>=1.24.0",
]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-q"
testpaths = ["tests"]
